requirements.txt
Flask==3.0.3
face-recognition==1.3.0
opencv-python==4.10.0.84
mysql-connector-python==9.1.0
Pillow==11.0.0
pandas==2.2.3
numpy==2.1.3
